+++ backend/graphql/units.resolvers.js
@@ -1,3 +1,4 @@
+const boom = require('@hapi/boom');
 const UnitsService = require('../services/units.service');
 const service = new UnitsService();
 
@@ -23,7 +24,13 @@ const unit = (_, { number }) => {
  * @param {Object} dto - Objeto con los datos de la unidad
  * @returns {Object} Objeto con la unidad
  */
-const createUnit = (_, { dto }) => {
+const createUnit = async (_, { dto }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+
+	if (!user) {
+		throw boom.unauthorized('No autorizado');
+	}
+
 	return service.create(dto);
 };
 
@@ -33,7 +40,13 @@ const createUnit = (_, { dto }) => {
  * @param {Object} dto - Objeto con los datos de la unidad
  * @returns {Object} Objeto con la unidad
  */
-const updateUnit = (_, { id, dto }) => {
+const updateUnit = async (_, { id, dto }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+
+	if (!user) {
+		throw boom.unauthorized('No autorizado');
+	}
+
 	return service.update(id, dto);
 };
 
@@ -42,7 +55,13 @@ const updateUnit = (_, { id, dto }) => {
  * @param {number} id - id de la unidad
  * @returns {number} id de la unidad eliminada
  */
-const deleteUnit = async (_, { id }) => {
+const deleteUnit = async (_, { id }, context) => {
+	const { user } = await context.authenticate('jwt', { session: false });
+
+	if (!user) {
+		throw boom.unauthorized('No autorizado');
+	}
+
 	await service.delete(id);
 	return id;
 };
